// src/Client/Logistics.Angular/projects/shared/src/lib/api/generated/models/tenant-settings.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import type { CurrencyCode } from '../models/currency-code';
import type { DateFormatType } from '../models/date-format-type';
import type { DistanceUnit } from '../models/distance-unit';
import type { WeightUnit } from '../models/weight-unit';
export interface TenantSettings {
  currency?: CurrencyCode;
  dateFormat?: DateFormatType;
  distanceUnit?: DistanceUnit;
  timezone?: string | null;
  weightUnit?: WeightUnit;
}